+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Expose `NormalizedBlendValues` from `VirtualBlendTree`
+- Added `IPlatformAnimatorBindings.PreCommitController`
 
 ### Fixed
 - [#675] Fixed comparison function in `SingleObjectQuery.ObserveTransformPosition`